fix: testResult 엔티티 필드 변경

+++ src/main/java/com/weflo/backend/domain/drone/domain/Drone.java
@@ -29,7 +29,7 @@ public class Drone {
 
     private String purpose;
 
-    private Long cost;
+    private int cost;
 
     private int flightCount;
